Commonize PAL logic between POSIX and WinSim of mapping to OS files
