+++ source/portable/os/freertos_winsim/core_pkcs11_pal.c
@@ -44,38 +44,8 @@
 #include <stdio.h>
 #include <string.h>
 
-/**
- * @ingroup pkcs11_macros
- * @brief Macros for managing PKCS #11 objects in flash.
- *
- */
-#define pkcs11palFILE_NAME_CLIENT_CERTIFICATE    "FreeRTOS_P11_Certificate.dat"       /**< The file name of the Certificate object. */
-#define pkcs11palFILE_NAME_PUBLIC_KEY            "FreeRTOS_P11_PubKey.dat"            /**< The file name of the Public Key object. */
-#define pkcs11palFILE_NAME_KEY                   "FreeRTOS_P11_Key.dat"               /**< The file name of the Private Key object. */
-#define pkcs11palFILE_CODE_SIGN_PUBLIC_KEY       "FreeRTOS_P11_CodeSignKey.dat"       /**< The file name of the Code Sign Key object. */
-#define pkcs11palFILE_CMAC_SECRET_KEY            "FreeRTOS_P11_CMACKey.dat"           /**< The file name of the CMAC Secret Key object. */
+#include "core_pkcs11_pal_utils.h"
 
-/**
- * @ingroup pkcs11_macros
- * @brief PKCS #11 logging macro.
- *
- */
-#define PKCS11_PAL_PRINT( X )    configPRINTF( X )
-
-/**
- * @ingroup pkcs11_enums
- * @brief Enums for managing PKCS #11 object types.
- *
- */
-enum eObjectHandles
-{
-    eInvalidHandle = 0,       /**< According to PKCS #11 spec, 0 is never a valid object handle. */
-    eAwsDevicePrivateKey = 1, /**< Private Key. */
-    eAwsDevicePublicKey,      /**< Public Key. */
-    eAwsDeviceCertificate,    /**< Certificate. */
-    eAwsCodeSigningKey,       /**< Code Signing Key. */
-    eAwsCMACSecretKey         /**< CMAC Secret Key. */
-};
 
 /*-----------------------------------------------------------*/
 
@@ -102,124 +72,6 @@ BaseType_t prvFileExists( const char * pcFileName )
     }
 }
 
-/**
- * @brief Maps label to filename and object handle.
- * @param[in] pcLabel            The PKCS #11 label to convert to a file name
- * @param[out] pcFileName        The name of the file to check for existence.
- * @param[out] pHandle           The type of the PKCS #11 object.
- *
- */
-void prvLabelToFilenameHandle( uint8_t * pcLabel,
-                               char ** pcFileName,
-                               CK_OBJECT_HANDLE_PTR pHandle )
-{
-    if( pcLabel != NULL )
-    {
-        /* Translate from the PKCS#11 label to local storage file name. */
-        if( 0 == memcmp( pcLabel,
-                         pkcs11configLABEL_DEVICE_CERTIFICATE_FOR_TLS,
-                         sizeof( pkcs11configLABEL_DEVICE_CERTIFICATE_FOR_TLS ) ) )
-        {
-            *pcFileName = pkcs11palFILE_NAME_CLIENT_CERTIFICATE;
-            *pHandle = eAwsDeviceCertificate;
-        }
-        else if( 0 == memcmp( pcLabel,
-                              pkcs11configLABEL_DEVICE_PRIVATE_KEY_FOR_TLS,
-                              sizeof( pkcs11configLABEL_DEVICE_PRIVATE_KEY_FOR_TLS ) ) )
-        {
-            *pcFileName = pkcs11palFILE_NAME_KEY;
-            *pHandle = eAwsDevicePrivateKey;
-        }
-        else if( 0 == memcmp( pcLabel,
-                              pkcs11configLABEL_DEVICE_PUBLIC_KEY_FOR_TLS,
-                              sizeof( pkcs11configLABEL_DEVICE_PUBLIC_KEY_FOR_TLS ) ) )
-        {
-            *pcFileName = pkcs11palFILE_NAME_PUBLIC_KEY;
-            *pHandle = eAwsDevicePublicKey;
-        }
-        else if( 0 == memcmp( pcLabel,
-                              pkcs11configLABEL_CODE_VERIFICATION_KEY,
-                              sizeof( pkcs11configLABEL_CODE_VERIFICATION_KEY ) ) )
-        {
-            *pcFileName = pkcs11palFILE_CODE_SIGN_PUBLIC_KEY;
-            *pHandle = eAwsCodeSigningKey;
-        }
-        else if( 0 == memcmp( pcLabel,
-                              pkcs11configLABEL_CMAC_KEY,
-                              sizeof( pkcs11configLABEL_CMAC_KEY ) ) )
-        {
-            *pcFileName = pkcs11palFILE_CMAC_SECRET_KEY;
-            *pHandle = ( CK_OBJECT_HANDLE ) eAwsCMACSecretKey;
-        }
-        else
-        {
-            *pcFileName = NULL;
-            *pHandle = eInvalidHandle;
-        }
-    }
-}
-
-/**
- * @brief Maps object handle to file name
- *
- * @param[in] pcLabel            The PKCS #11 label to convert to a file name
- * @param[out] pcFileName        The name of the file to check for existence.
- * @param[out] pHandle           The type of the PKCS #11 object.
- *
- */
-static CK_RV prvHandleToFilename( CK_OBJECT_HANDLE xHandle,
-                                  const char ** pcFileName,
-                                  CK_BBOOL * pIsPrivate )
-{
-    CK_RV xReturn = CKR_OK;
-
-    if( pcFileName != NULL )
-    {
-        switch( ( CK_OBJECT_HANDLE ) xHandle )
-        {
-            case eAwsDeviceCertificate:
-                *pcFileName = pkcs11palFILE_NAME_CLIENT_CERTIFICATE;
-                /* coverity[misra_c_2012_rule_10_5_violation] */
-                *pIsPrivate = ( CK_BBOOL ) CK_FALSE;
-                break;
-
-            case eAwsDevicePrivateKey:
-                *pcFileName = pkcs11palFILE_NAME_KEY;
-                /* coverity[misra_c_2012_rule_10_5_violation] */
-                *pIsPrivate = ( CK_BBOOL ) CK_TRUE;
-                break;
-
-            case eAwsDevicePublicKey:
-                *pcFileName = pkcs11palFILE_NAME_PUBLIC_KEY;
-                /* coverity[misra_c_2012_rule_10_5_violation] */
-                *pIsPrivate = ( CK_BBOOL ) CK_FALSE;
-                break;
-
-            case eAwsCodeSigningKey:
-                *pcFileName = pkcs11palFILE_CODE_SIGN_PUBLIC_KEY;
-                /* coverity[misra_c_2012_rule_10_5_violation] */
-                *pIsPrivate = ( CK_BBOOL ) CK_FALSE;
-                break;
-
-            case eAwsCMACSecretKey:
-                *pcFileName = pkcs11palFILE_CMAC_SECRET_KEY;
-                /* coverity[misra_c_2012_rule_10_5_violation] */
-                *pIsPrivate = ( CK_BBOOL ) CK_TRUE;
-                break;
-
-            default:
-                xReturn = CKR_KEY_HANDLE_INVALID;
-                break;
-        }
-    }
-    else
-    {
-        LogError( ( "Could not convert label to filename. Received a NULL parameter." ) );
-    }
-
-    return xReturn;
-}
-
 /*-----------------------------------------------------------*/
 
 CK_RV PKCS11_PAL_Initialize( void )
@@ -238,9 +90,9 @@ CK_OBJECT_HANDLE PKCS11_PAL_SaveObject( CK_ATTRIBUTE_PTR pxLabel,
     CK_OBJECT_HANDLE xHandle = eInvalidHandle;
 
     /* Converts a label to its respective filename and handle. */
-    prvLabelToFilenameHandle( pxLabel->pValue,
-                              &pcFileName,
-                              &xHandle );
+    PAL_UTILS_LabelToFilenameHandle( pxLabel->pValue,
+                                     &pcFileName,
+                                     &xHandle );
 
     /* If your project requires additional PKCS#11 objects, add them here. */
 
@@ -258,7 +110,7 @@ CK_OBJECT_HANDLE PKCS11_PAL_SaveObject( CK_ATTRIBUTE_PTR pxLabel,
         if( INVALID_HANDLE_VALUE == hFile )
         {
             ulStatus = GetLastError();
-            PKCS11_PAL_PRINT( ( "ERROR: Unable to create file %d \r\n", ulStatus ) );
+            LogError( ( "Unable to create file %d \r\n", ulStatus ) );
             xHandle = eInvalidHandle;
         }
 
@@ -295,9 +147,9 @@ CK_OBJECT_HANDLE PKCS11_PAL_FindObject( CK_BYTE_PTR pxLabel,
     char * pcFileName = NULL;
 
     /* Converts a label to its respective filename and handle. */
-    prvLabelToFilenameHandle( pxLabel,
-                              &pcFileName,
-                              &xHandle );
+    PAL_UTILS_LabelToFilenameHandle( pxLabel,
+                                     &pcFileName,
+                                     &xHandle );
 
     /* Check if object exists/has been created before returning. */
     if( pdTRUE != prvFileExists( pcFileName ) )
@@ -321,9 +173,9 @@ CK_RV PKCS11_PAL_GetObjectValue( CK_OBJECT_HANDLE xHandle,
     char * pcFileName = NULL;
 
 
-    ulReturn = prvHandleToFilename( xHandle,
-                                    &pcFileName,
-                                    pIsPrivate );
+    ulReturn = PAL_UTILS_HandleToFilename( xHandle,
+                                           &pcFileName,
+                                           pIsPrivate );
 
     if( ( pcFileName != NULL ) && ( pdTRUE == prvFileExists( pcFileName ) ) )
     {
@@ -339,7 +191,7 @@ CK_RV PKCS11_PAL_GetObjectValue( CK_OBJECT_HANDLE xHandle,
         if( INVALID_HANDLE_VALUE == hFile )
         {
             ulDriverReturn = GetLastError();
-            PKCS11_PAL_PRINT( ( "ERROR: Unable to open file %d \r\n", ulDriverReturn ) );
+            LogError( ( "Unable to open file %d \r\n", ulDriverReturn ) );
             ulReturn = CKR_FUNCTION_FAILED;
         }
 
@@ -366,7 +218,7 @@ CK_RV PKCS11_PAL_GetObjectValue( CK_OBJECT_HANDLE xHandle,
                                    ( LPDWORD ) ( &ulSize ),
                                    NULL ) )
             {
-                PKCS11_PAL_PRINT( ( "ERROR: Unable to read file \r\n" ) );
+                LogError( ( "Unable to read file \r\n" ) );
                 ulReturn = CKR_FUNCTION_FAILED;
             }
         }
